Fix off-by-one error in postprocessing.
Thanks @sztaylor for reporting.

For the first 500 images in the ImageNet validation set, "accuracy_topn" was
"yes" for 450 images, while "accuracy_top5" was "yes" only for 440 images. For
all mismatched images, a pattern emerged as in the following example:

  "ILSVRC2012_val_00000334.JPEG": {
    "accuracy_top5": "no",
    "accuracy_top1": "no",
    "accuracy_topn": "yes",
    "class_correct": 733,
    "class_topmost": 616
    "topn": [
      ...
      {
        "index": 733,
        "name": "(733) n03976657 pole",
        "prob": 0.0410714
      }
    ],
  }

Namely, the image label ("class_correct" : 733) matched the fifth most
probable result ("index" : 733). According to the definition,
"accuracy_topn", where n=5, was "yes". Unfortunately, "accuracy_top5"
was incorrectly set to "no".

The bug was in the postprocesing script, where "accuracy_top5" was
calculated by the Boolean expression "class_correct in classes[0:4]".
Since in Python ranges exclude the right bound, "accuracy_top5"
in fact corresponded to "accuracy_top4".